+++ Domain/Clients/Models/Client.php
@@ -7,7 +7,7 @@ namespace Domain\Clients\Models;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
-use App\Item;
+use Domain\Items\Models\Item;
 
 class Client extends Model
 {
