+++ Domain/Items/Models/Item.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace Domain\Items\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
